+++ playwright.config.ts
@@ -11,7 +11,9 @@ dotenv.config({ path: ".env.test" });
  * See https://playwright.dev/docs/test-configuration.
  */
 export default defineConfig({
-  testDir: "./e2e",
+  testDir: ".",
+  testMatch: "**/*.spec.ts",
+  testIgnore: "**/node_modules/**/*",
   /* Run tests in files in parallel */
   fullyParallel: true,
   /* Fail the build on CI if you accidentally left test.only in the source code. */
@@ -77,5 +79,5 @@ export default defineConfig({
     reuseExistingServer: !process.env.CI,
   },
 
-  globalSetup: require.resolve("./e2e/global-setup"),
+  globalSetup: require.resolve("./src/mocks/global-setup"),
 });
