+++ src/app/fortune/[facultyCategoryPathName]/components/__tests__/FacultyFortuneModal.test.tsx
@@ -1,9 +1,9 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import "@testing-library/jest-dom";
-import FacultyFortuneModal from "@/components/FacultyFortuneModal";
+import FacultyFortuneModal from "@/app/fortune/[facultyCategoryPathName]/components/FacultyFortuneModal";
 import * as FacultyFortuneResultFetchers from "@/lib/fetchFacultyFortuneResult";
-import { postOffice } from "../fixtures/post_offices";
+import { postOffice } from "@/fixtures/post_offices";
 
 const user = userEvent.setup();
