+++ src/mocks/mockVillageFortune.ts
